xrandr: Use floating point for VTotal when calculating refresh rate
Interlaced modes generally have an odd VTotal, so we lose half a line
from VTotal when we divide by two. That causes the final refresh rate
to be slightly off. Make VTotal a double to avoid the problem.

+++ xrandr.c
@@ -543,7 +543,7 @@ static double
 mode_refresh (XRRModeInfo *mode_info)
 {
     double rate;
-    unsigned int vTotal = mode_info->vTotal;
+    double vTotal = mode_info->vTotal;
 
     if (mode_info->modeFlags & RR_DoubleScan) {
 	/* doublescan doubles the number of lines */
